Trait inscribe::rankers::Ranker

pub trait Ranker {
    fn rank(
        self, 
        results: Vec<&mut Record>, 
        keyword: String
    ) -> Result<Vec<(u32, String)>, Box<dyn Error + Send + Sync>>;
}

A Ranker interface/trait, it's passed the DB Record and returns a 'rank' sorted list of same no traits for collections yet in rust, so this is verbose to abstract

Required methods

fn rank(
self,
results: Vec<&mut Record>,
keyword: String
) -> Result<Vec<(u32, String)>, Box<dyn Error + Send + Sync>>

Loading content...